+++ data/tasks.txt
@@ -1,3 +1,5 @@
 [T][✘] hw
 [T][✓] gym
-[T][✘] sleep
+[T][✘] 1
+[D][✓] 2 (by: 31/8/2019 1200)
+[E][✘] (at: 20/11/2020 1400-2100)
