--- src/main/java/Border.java
@@ -1,9 +1,0 @@
-public class Border {
-    public String toString(){
-        String output = "    ";
-        for (int n = 0 ; n < 60 ; n ++){
-            output += "_";
-        }
-        return output;
-    }
-}
