+++ src/main/java/Duke.java
@@ -1,4 +1,10 @@
-import java.util.Scanner;
+import command.Command;
+import exception.DukeException;
+import filewriter.Storage;
+import parser.Parser;
+import task.TaskList;
+import ui.Ui;
+
 
 public class Duke {
 
@@ -32,6 +38,7 @@ public class Duke {
             }
         }
         ui.showGoodbye();
+
     }
 
     public static void main(String[] args) {
