--- src/main/java/ExitCommand.java
@@ -1,9 +1,0 @@
-public class ExitCommand extends Command {
-    public boolean isExit(){
-        return true;
-    }
-
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException{
-        storage.closeWriter();
-    };
-}
