+++ src/main/java/command/Command.java
@@ -1,3 +1,10 @@
+package command;
+
+import exception.DukeException;
+import filewriter.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public abstract class Command {
     FullCommand type;
