+++ src/main/java/command/EditCommand.java
@@ -1,3 +1,11 @@
+package command;
+
+import exception.DukeException;
+import filewriter.Storage;
+import task.Task;
+import task.TaskList;
+import ui.Ui;
+
 public class EditCommand extends Command {
     int index;
 
@@ -9,7 +17,7 @@ public class EditCommand extends Command {
         this.index = index;
     }
 
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException{
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         try {
             if (index == -1){
                 throw new IndexOutOfBoundsException();
