+++ src/main/java/command/ExitCommand.java
@@ -1,0 +1,16 @@
+package command;
+
+import exception.DukeException;
+import filewriter.Storage;
+import task.TaskList;
+import ui.Ui;
+
+public class ExitCommand extends Command {
+    public boolean isExit(){
+        return true;
+    }
+
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        storage.closeWriter();
+    };
+}
