+++ src/main/java/command/ListCommand.java
@@ -1,0 +1,16 @@
+package command;
+
+import exception.DukeException;
+import filewriter.Storage;
+import task.TaskList;
+import ui.Ui;
+
+public class ListCommand extends Command {
+    public boolean isExit(){
+        return false;
+    }
+
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        ui.readList(tasks);
+    };
+}
