+++ src/main/java/task/Event.java
@@ -1,3 +1,8 @@
+package task;
+
+import datetime.DateTime;
+import exception.DukeException;
+
 public class Event extends Task{
     DateTime date_Time;
     public Event(String description) throws DukeException{
