+++ src/main/java/task/Todo.java
@@ -1,5 +1,9 @@
+package task;
+
+import exception.DukeException;
+
 public class Todo extends Task{
-    public Todo(String description) throws DukeException{
+    public Todo(String description) throws DukeException {
         super(description);
     }
